+++ packages/components/src/buttons/IconLabeledButton.js
@@ -72,7 +72,7 @@ const IconLabelledBaseButton = addButtonStyles(
  *
  * @returns {ReactElement} Styled icon button.
  */
-const IconLabelledButton = ( props ) => {
+const IconLabeledButton = ( props ) => {
 	const { children, icon, textColor } = props;
 
 	const newProps = omit( props, "icon" );
@@ -85,7 +85,7 @@ const IconLabelledButton = ( props ) => {
 	);
 };
 
-IconLabelledButton.propTypes = {
+IconLabeledButton.propTypes = {
 	type: PropTypes.string,
 	icon: PropTypes.string.isRequired,
 	textColor: PropTypes.string,
@@ -108,7 +108,7 @@ IconLabelledButton.propTypes = {
 	] ).isRequired,
 };
 
-IconLabelledButton.defaultProps = {
+IconLabeledButton.defaultProps = {
 	type: "button",
 	textColor: colors.$color_blue,
 	textFontSize: "inherit",
@@ -125,4 +125,4 @@ IconLabelledButton.defaultProps = {
 	focusBorderColor: colors.$color_blue,
 };
 
-export default IconLabelledButton;
+export default IconLabeledButton;
