+++ packages/search-metadata-previews/src/snippet-editor/SnippetEditor.js
@@ -6,9 +6,8 @@ import noop from "lodash/noop";
 
 // Yoast dependencies.
 import { assessments, string, helpers } from "yoastseo";
-import { Button } from "yoast-components/composites/Plugin/Shared/components/Button";
 import { colors } from "@yoast/components/style-guide";
-import { ErrorBoundary, SvgIcon } from "@yoast/components";
+import { ErrorBoundary, SvgIcon, Button } from "@yoast/components";
 import { getDirectionalStyle } from "@yoast/helpers";
 
 const { MetaDescriptionLengthAssessment, PageTitleWidthAssessment } = assessments.seo;
