+++ packages/yoast-components/composites/Plugin/Shared/tests/LinkButtonTest.js
@@ -1,7 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 
-import { BaseLinkButton, LinkButton } from "../components/LinkButton";
+import { BaseLinkButton, LinkButton } from "../../../../../components/src/buttons/LinkButton";
 
 test( "the BaseLinkButton matches the snapshot", () => {
 	const component = renderer.create(
